requirements.txt

flask
binance-futures-connector
